+++ docs/pages/_en/1.0/reference/release-notes.md
@@ -15,6 +15,7 @@ group: Deployment & Configs
 - Added 'Query Namesake Element' component.
 - Now 'Query Elements' give elements in reverse order when 'Limit' is negative.
 - Improved 'Host Boundary Profile' to accept a plane as an input.
+- Fix on '"Add Model Element' for slanted walls.
 
 ### RC